+++ inst/resources/init-rprofile.R
@@ -1,3 +1,3 @@
-#### -- Packrat Autoloader (version 0.5.0-1) -- ####
+#### -- Packrat Autoloader (version 0.5.0-2) -- ####
 source("packrat/init.R")
 #### -- End Packrat Autoloader -- ####